requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
google_search_results==1.7.1
pyTelegramBotAPI==2.1.7
requests==2.22.0
lyricsgenius==1.8.2